Amazon Kinesis/Python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paho-mqtt==1.6.1
boto3==1.35.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
pydantic==2.5.0
python-multipart==0.0.6